# optimize clvm interpreter (benchmarks -18.3%)