This attempts to address recent performance regression test failures, most notably the ones currently blocking: https://github.com/Chia-Network/chia_rs/pull/68

The benchmarks improvements are as follows:

```
      block-2000.hex mean: 0.017147 (-0.004450) -20.60 %
 compressed-2000.hex mean: 1.496101 (-0.577866) -27.86 %
          concat.hex mean: 0.680109 (-0.004657) -0.68 % (within uncertainty)
      count-even.hex mean: 0.031021 (-0.007084) -18.59 %
       factorial.hex mean: 0.259588 (+0.008307) +3.31 % (within uncertainty)
     hash-string.hex mean: 0.033362 (+0.001150) +3.57 % (within uncertainty)
       hash-tree.hex mean: 0.081733 (-0.020562) -20.10 %
     large-block.hex mean: 0.314545 (-0.086245) -21.52 %
        loop_add.hex mean: 6.015117 (-1.403589) -18.92 %
        loop_ior.hex mean: 5.878794 (-1.726914) -22.71 %
        loop_not.hex mean: 5.538171 (-1.305278) -19.07 %
        loop_sub.hex mean: 5.985422 (-1.548787) -20.56 %
        loop_xor.hex mean: 5.893141 (-1.570511) -21.04 %
 matrix-multiply.hex mean: 0.355724 (-0.101242) -22.16 %
       point-pow.hex mean: 0.434597 (-0.007322) -1.66 % (within uncertainty)
     pubkey-tree.hex mean: 0.532662 (-0.019352) -3.51 % (within uncertainty)
      shift-left.hex mean: 3.590006 (+0.076966) +2.19 % (within uncertainty)
     substr-tree.hex mean: 0.280550 (-0.057000) -16.89 %
          substr.hex mean: 0.013477 (-0.002452) -15.39 %
        sum-tree.hex mean: 0.632173 (-0.167014) -20.90 %
TOTAL: 190.317203 (-42.619505) -18.30 %
```

Some of the `run-programs.py` tests also show performance improvements. The omitted ones are unchanged:

```
/usr/bin/time ./run.py programs/args-unknown-5.hex programs/args-unknown-5.envhex
  Runtime: 4.74 s (down from 5.91 s)
  Resident Size: 541 MiB (down from 1005 MB)
/usr/bin/time ./run.py programs/args-unknown-6.hex programs/args-unknown-6.envhex
  Runtime: 9.69 s (down from 12.88 s)
  Resident Size: 1142 MiB (down from 2140 MB)
/usr/bin/time ./run.py programs/args-unknown-7.hex programs/args-unknown-7.envhex
  Runtime: 7.10 s (down from 9.00 s)
  Resident Size: 806 MiB (down from 1504)
/usr/bin/time ./run.py programs/args-unknown-8.hex programs/args-unknown-8.envhex
  Runtime: 6.29 s (down from 8.13 s)
  Resident Size: 733 MiB (down from 1366 MB)
/usr/bin/time ./run.py programs/args-unknown-9.hex programs/args-unknown-9.envhex
  Runtime: 8.19 s (down from 10.42 s)
  Resident Size: 922 MiB (down from 1724 MB)
/usr/bin/time ./run.py programs/recursive-add.hex programs/recursive-add.envhex
  Runtime: 2.87 s (down from 2.92)
  Resident Size: 684 MiB (down from 700 MB)
/usr/bin/time ./run.py programs/recursive-cons.hex programs/recursive-cons.envhex
  Runtime: 9.47 s (down from 12.29 s)
  Resident Size: 1239 MiB (down from 2269 MB)
/usr/bin/time ./run.py programs/recursive-sub.hex programs/recursive-sub.envhex
  Runtime: 7.28 s (down from 9.28 s)
  Resident Size: 837 MiB (down from 1537 MB)
```